backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
Pillow>=10.0
python-decouple>=3.8
gunicorn>=21.2
psycopg2-binary>=2.9
